requirements.txt
Django==3.1.1
django-storages==1.10.1
drf-yasg[validation]==1.17.1
django-prometheus==2.0.0 
django-tables2==2.3.1
djangorestframework==3.11.2
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-filter==2.4.0
django-timezone-field==4.0
psycopg2-binary==2.8.5
Markdown==3.2.2
PyYAML==5.3.1
django-taggit==1.3.0
netaddr==0.8.0
boto3==1.12.39

